Implementacion de view y view controller de home, pintado de data de usuario y ordenamiento de codigo
