+++ src/index.html
@@ -35,12 +35,6 @@
   firebase.initializeApp(firebaseConfig);
 </script>
  
-
-
-
-
-
-
   <script type="module" src="./main.js"></script>
 </body>
