+++ src/main.js
@@ -1,4 +1,4 @@
-import { changeView } from './view-controller/route.js';
+import { changeView } from './view-controller/route-controller.js';
 
 const init = () => {
   // Para que se haga un cambio cuando se recarga la página
